Update
Added homeworld traits, career starting skills, talents and traits.

All binary choice skills and talents are picked randomly with 50% weight
to either choice - it simulates a 1d10 roll per option.

Changed input parameters to simplify use. Removed Gender and Race. This
can be elaborated on by the player in the biography if needs be.